add content script for new reddit redesign
Making this work on the new reddit redesigned site is proving to be
tricky due to the single page application aspect. The script works when
a post is opened up in overlay / modal form. It does not handle when the
page is initially loaded on the comments page.

To work around the fact a SPA doesn't reload the page, the script
listens to the click event paired with a request animation frame and
compares a stored value of the url pathname to the current pathname.
Maybe there is a better way to do this.

Browsing with a mouse works, but keyboard navigation is currently not handled.

There is an edge case where clicking outside the post overlay to close
and go back to browsing all posts doesn't seem to give a click event.
Since there is no click event, the stored url pathname doesn't change,
so if the same post is then opened again, the pathname will be the same,
and therefore it won't check the link.